Eliminated compiler warnings for ManualArm (read description)
The reason we were getting the warnings was that switches only support
"integral types", which include all variants of integers (int,
short,long). nNxtButtonPressed technically does not have an integral
type. Rather, the type is TNxtButtons, which to us looks like an
integer, but it's not really.
The change I made to the code is called type casting. We're essentially
telling the compiler to convert the variable's type for use in this one
instance. All you have to do is put the desired type in parenthesis next
to the variable that you want converted, like I did here. If you need a
whole expression converted to a different type, you use parenthesis
around the expression in addition to around the desired type.