adjusted #pragma and removed unused code
since we got rid of the extra motor controller, I changed the pragma so
the arm motor is in the first MC with the harvester and the servo
controller is now third in the daisy chain.  I removed the old lines of
code controlling two Lego motors now that we have the quarter scale
servo.